moved model util, fix tests
